renaming to main & improving
